requirements.txt
Django==4.2.7
mysqlclient==2.2.0
django-crispy-forms==2.0
Pillow==10.0.0
python-dotenv==1.0.0
crispy-bootstrap5==2024.10